requirements.txt
enum34==1.1.6
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Injector==0.10.1
Flask-Script==2.0.6
PyJWT==1.5.3
